Connect 4 Game
Connect 4 is a two-player connection board game, in which the players choose a color and then take turns dropping colored discs into a seven-column, six-row vertically suspended grid.

The pieces fall straight down, occupying the lowest available space within the column.

The objective of the game is to be the first to form a horizontal, vertical, or diagonal line of four of one’s own discs.